refactor: extract runtime read models
